+++ src/Mailman.Server/ClientApp/src/actions/CreateMergeTemplate.js
