+++ src/Mailman.Server/ClientApp/src/components/CreateMerge/TabSelectionContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 
 import TabSelection from './TabSelection'
-import { updateTabSelection } from '../../actions/createMergeTemplate'
+import { updateTabSelection } from '../../actions/CreateMergeTemplate'
 
 function mapStateToProps(state, ownProps) {
   const { currentMergeTemplate, mergeTemplates } = state // Get this from the store
