+++ src/Mailman.Server/ClientApp/src/index.js
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'react-router-redux'
 import { createBrowserHistory } from 'history'
 import { MuiThemeProvider } from '@material-ui/core/styles'
-import configureStore from './store/configureStore'
+import configureStore from './store/ConfigureStore'
 import App from './App'
 import registerServiceWorker from './registerServiceWorker'
 import theme from './theme'
