+++ src/Mailman.Server/ClientApp/src/store/configureStore.js
@@ -5,7 +5,7 @@ import * as Counter from '../reducers/Counter'
 import * as WeatherForecasts from '../reducers/WeatherForecasts'
 import * as NavDrawer from '../reducers/NavDrawer'
 import * as MergeTemplates from '../reducers/readMergeTemplates'
-import { currentMergeTemplateReducer } from '../reducers/createMergeTemplate'
+import { currentMergeTemplateReducer } from '../reducers/CreateMergeTemplate'
 
 export default function configureStore(history, initialState) {
   const reducers = {
